+++ demisto_sdk/commands/common/errors.py
@@ -9,7 +9,8 @@ from demisto_sdk.commands.common.constants import (BETA_INTEGRATION_DISCLAIMER,
 FOUND_FILES_AND_ERRORS: list = []
 FOUND_FILES_AND_IGNORED_ERRORS: list = []
 
-ALLOWED_IGNORE_ERRORS = ['BA101', 'BA106', 'RP102', 'RP104', 'SC100', 'IF106', 'PA113', 'PA116', 'IN126', 'PB105']
+ALLOWED_IGNORE_ERRORS = ['BA101', 'BA106', 'RP102', 'RP104', 'SC100', 'IF106', 'PA113', 'PA116', 'IN126', 'PB105',
+                         'PB106']
 
 PRESET_ERROR_TO_IGNORE = {
     'community': ['BC', 'CJ', 'DS', 'PA117', 'IN125', 'IN126'],
@@ -110,6 +111,7 @@ ERROR_CODE = {
     "playbook_unconnected_tasks": "PB103",
     "invalid_deprecated_playbook": "PB104",
     "playbook_cant_have_deletecontext_all": "PB105",
+    "using_instance_in_playbook": "PB106",
     "description_missing_in_beta_integration": "DS100",
     "no_beta_disclaimer_in_description": "DS101",
     "no_beta_disclaimer_in_yml": "DS102",
@@ -714,6 +716,11 @@ class Errors:
     @error_code_decorator
     def playbook_cant_have_rolename():
         return "Playbook can not have a rolename."
+
+    @staticmethod
+    @error_code_decorator
+    def using_instance_in_playbook():
+        return "Playbook should not use specific instance."
 
     @staticmethod
     @error_code_decorator