+++ demisto_sdk/commands/common/hook_validations/playbook.py
@@ -30,6 +30,7 @@ class PlaybookValidator(ContentEntityValidator):
                 self.is_id_equals_name(),
                 self.is_no_rolename(),
                 self.is_root_connected_to_all_tasks(),
+                self.is_using_instance(),
                 self.is_condition_branches_handled(),
                 self.is_delete_context_all_in_playbook(),
                 self.are_tests_configured(),
@@ -43,6 +44,7 @@ class PlaybookValidator(ContentEntityValidator):
                 self.is_valid_version(),
                 self.is_no_rolename(),
                 self.is_root_connected_to_all_tasks(),
+                self.is_using_instance(),
                 self.is_condition_branches_handled(),
                 self.is_delete_context_all_in_playbook(),
                 self.are_tests_configured()
@@ -269,6 +271,22 @@ class PlaybookValidator(ContentEntityValidator):
             if curr_task and scriptargs and curr_task.get('scriptName', '') == 'DeleteContext' \
                     and scriptargs.get('all', {}).get('simple', '') == 'yes':
                 error_message, error_code = Errors.playbook_cant_have_deletecontext_all()
+                if self.handle_error(error_message, error_code, file_path=self.file_path):
+                    self.is_valid = False
+                    return False
+        return True
+
+    def is_using_instance(self) -> bool:
+        """
+        Check if there is an existing task that uses specific instance.
+        Returns:
+            True if using specific instance exists else False.
+        """
+        tasks: Dict = self.current_file.get('tasks', {})
+        for task in tasks.values():
+            scriptargs = task.get('scriptarguments', {})
+            if scriptargs and scriptargs.get('using', {}):
+                error_message, error_code = Errors.using_instance_in_playbook()
                 if self.handle_error(error_message, error_code, file_path=self.file_path):
                     self.is_valid = False
                     return False